Remove the GoogleResponseHandler class
